refactor: changed schema models naming format
